+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+#### ⚙️ Internal
+
+- Updated proto to [v0.46.0](https://github.com/moonrepo/proto/releases/tag/v0.46.0) (from 0.45.2).
+
 ## 1.32.5
 
 #### 🚀 Updates